Separate instance from main.go file
